+++ nars_gui/src/main/java/automenta/vivisect/swing/NSliderSwing.java
@@ -45,7 +45,7 @@ public class NSliderSwing extends JLabel implements MouseListener, MouseMotionLi
     protected NumberFormat nf = NumberFormat.getInstance();
     protected String prefix = "";
 
-    final static Font defaultLabelFont = Video.monofont.deriveFont(11f);
+    final static Font defaultLabelFont = Video.monofont.deriveFont(11.0f);
 
     public NSliderSwing() {
         this(0,0,0);
@@ -98,7 +98,7 @@ public class NSliderSwing extends JLabel implements MouseListener, MouseMotionLi
         float p = (value.floatValue() - min) / (max-min);
         if (barColor == null) {
             //Green->Yellow->Red
-            g.setColor(Color.getHSBColor( (1f - (float)p) / 2.5f + 0.15f ,
+            g.setColor(Color.getHSBColor( (1.0f - (float)p) / 2.5f + 0.15f ,
                     (0.2f *((float)p)) + 0.7f,
                     (0.2f *((float)p)) + 0.7f));
             // g.setColor(Color.getHSBColor( (1f - p) / 3.0f , 0.2f, 0.8f + 0.15f));
@@ -107,7 +107,7 @@ public class NSliderSwing extends JLabel implements MouseListener, MouseMotionLi
             g.setColor(barColor);
         }
         
-        int wp = (int)(((float) w) * p);
+        int wp = (int)((w) * p);
         g.fillRect(0, 0, wp, h);
 
 
@@ -139,7 +139,7 @@ public class NSliderSwing extends JLabel implements MouseListener, MouseMotionLi
     
 
     protected void updatePosition(int x) {
-        float p = ((float)x) / ((float)getWidth());
+        float p = ((float)x) / (getWidth());
         float v = p * (max-min) + min;
         v = Math.max(v, min);
         v = Math.min(v, max);
