+++ nars_lab/src/main/java/jurls/core/approximation/CNFBooleanFunction.java
@@ -64,7 +64,7 @@ public class CNFBooleanFunction implements ParameterizedFunction {
         }
 
         parameters = new long[cnf.length];
-        Arrays.fill(parameters, ~0l);
+        Arrays.fill(parameters, ~0L);
         intermediates = new long[cnf.length];
     }
 
@@ -117,7 +117,7 @@ public class CNFBooleanFunction implements ParameterizedFunction {
     }
 
     private long compute() {
-        long a = ~0l;
+        long a = ~0L;
 
         for (int j = 0; j < cnf.length; ++j) {
             long b = compute(j);
@@ -125,7 +125,7 @@ public class CNFBooleanFunction implements ParameterizedFunction {
             a &= b;
         }
 
-        a &= (1l << numOutputBits) - 1l;
+        a &= (1L << numOutputBits) - 1L;
         return a;
     }
 
@@ -134,24 +134,24 @@ public class CNFBooleanFunction implements ParameterizedFunction {
         int j = 0;
 
         for (int i = 0; i < xs.length; ++i) {
-            long v = Math.round(((1l << numBitsPerVariable[i]) - 1) * xs[i]);
+            long v = Math.round(((1L << numBitsPerVariable[i]) - 1) * xs[i]);
 
             for (int k = 0; k < numBitsPerVariable[i]; ++k, ++j) {
                 if (((v >> k) & 1) == 1) {
-                    variables[j] = ~0l;
+                    variables[j] = ~0L;
                 } else {
-                    variables[j] = 0l;
+                    variables[j] = 0L;
                 }
             }
         }
 
-        return (double) compute() / (double) ((1l << numOutputBits) - 1);
+        return (double) compute() / ((1L << numOutputBits) - 1);
     }
 
     @Override
     public void learn(double[] xs, double y) {
-        long currents = Math.round(value(xs) * ((1l << numOutputBits) - 1));
-        long targets = Math.round(y * ((1l << numOutputBits) - 1));
+        long currents = Math.round(value(xs) * ((1L << numOutputBits) - 1));
+        long targets = Math.round(y * ((1L << numOutputBits) - 1));
 
         final ArrayList<Integer> ps = new ArrayList<>(numOutputBits);
         for (int i = 0; i < numOutputBits; ++i) {
@@ -166,18 +166,18 @@ public class CNFBooleanFunction implements ParameterizedFunction {
                 }
             } else if (!target && current) {
                 for (int j = 0; j < parameters.length; ++j) {
-                    parameters[j] ^= 1l << i;
+                    parameters[j] ^= 1L << i;
                     if (((compute(j) >> i) & 1) == 0) {
                         ps.add(j);
                     }
-                    parameters[j] ^= 1l << i;
+                    parameters[j] ^= 1L << i;
                 }
             }
 
             // ps is the "LOGIC GRADIENT" (my invention)
             if (!ps.isEmpty()) {
                 int p = ps.get(random.nextInt(ps.size()));
-                parameters[p] ^= 1l << i;
+                parameters[p] ^= 1L << i;
             }
         }
     }
