+++ nars_logic/src/main/java/nars/Op.java
@@ -186,10 +186,10 @@ public enum Op implements Serializable {
             if (p < 31) //do not assign if it's an ordinary non-control char
                 this.byt = (byte) (p);
             else
-                this.byt = (byte) 0;
+                this.byt = 0;
         } else {
             //multiple ibytes, use the provided array
-            this.byt = (byte) 0;
+            this.byt = 0;
         }
 
         this.minLevel = minLevel;
