+++ nars_util/src/main/java/nars/util/data/bloom/Murmur3Hash.java
@@ -262,7 +262,7 @@ public class Murmur3Hash {
             case 10:
                 k2 ^= (long) (data[tailStart + 9] & 0xff) << 8;
             case 9:
-                k2 ^= (long) (data[tailStart + 8] & 0xff);
+                k2 ^= (data[tailStart + 8] & 0xff);
                 k2 *= C2;
                 k2 = Long.rotateLeft(k2, R3);
                 k2 *= C1;
@@ -283,7 +283,7 @@ public class Murmur3Hash {
             case 2:
                 k1 ^= (long) (data[tailStart + 1] & 0xff) << 8;
             case 1:
-                k1 ^= (long) (data[tailStart] & 0xff);
+                k1 ^= (data[tailStart] & 0xff);
                 k1 *= C1;
                 k1 = Long.rotateLeft(k1, R1);
                 k1 *= C2;
